requirements.txt
streamlit
langgraph
langchain-groq
langchain-community
langchain-experimental
pandas
plotly
scikit-learn
mysql-connector-python
python-dotenv
tabulate
openpyxl
sqlalchemy
psycopg2-binary
